create PostFixtures
