+++ src/DataFixtures/PostFixtures.php
@@ -1,0 +1,88 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\DataFixtures;
+
+use App\Entity\Comment;
+use App\Entity\Post;
+use Doctrine\Common\DataFixtures\DependentFixtureInterface;
+use Doctrine\Common\Persistence\ObjectManager;
+use function Symfony\Component\String\u;
+
+class PostFixtures extends AppFixtures implements DependentFixtureInterface
+{
+    public function load(ObjectManager $manager): void
+    {
+        foreach ($this->getPostData() as [$title, $slug, $summary, $content, $publishedAt, $author, $tags]) {
+            $post = new Post();
+            $post->setTitle($title);
+            $post->setSlug($slug);
+            $post->setSummary($summary);
+            $post->setContent($content);
+            $post->setPublishedAt($publishedAt);
+            $post->setAuthor($author);
+            $post->addTag(...$tags);
+
+            foreach (range(1, 5) as $i) {
+                $comment = new Comment();
+                $comment->setAuthor($this->getReference('john_user'));
+                $comment->setContent($this->getRandomText(random_int(255, 512)));
+                $comment->setPublishedAt(new \DateTime('now + '.$i.'seconds'));
+
+                $post->addComment($comment);
+            }
+
+            $manager->persist($post);
+        }
+
+        $manager->flush();
+    }
+
+    private function getPostData()
+    {
+        $posts = [];
+        foreach ($this->getPhrases() as $i => $title) {
+            // $postData = [$title, $slug, $summary, $content, $publishedAt, $author, $tags, $comments];
+            $posts[] = [
+                $title,
+                $this->slugger->slug($title)->lower(),
+                $this->getRandomText(),
+                $this->getPostContent(),
+                new \DateTime('now - '.$i.'days'),
+                // Ensure that the first post is written by Jane Doe to simplify tests
+                $this->getReference(['jane_admin', 'tom_admin'][0 === $i ? 0 : random_int(0, 1)]),
+                $this->getRandomTags(),
+            ];
+        }
+
+        return $posts;
+    }
+
+    private function getRandomText(int $maxLength = 255): string
+    {
+        $phrases = $this->getPhrases();
+        shuffle($phrases);
+
+        do {
+            $text = u('. ')->join($phrases)->append('.');
+            array_pop($phrases);
+        } while ($text->length() > $maxLength);
+
+        return $text;
+    }
+
+    public function getDependencies()
+    {
+        return [
+            UserFixtures::class
+        ];
+    }
+}
